Adds `Symbol` impls for `Display`, `Ord`, `PartialOrd`
Moves `Symbol` to its own module and adds implementations of the
following traits:
* `Display`
* `PartialOrd`
* `Ord`
