+++ src/lib.rs
@@ -28,19 +28,17 @@ mod raw_symbol_token;
 mod raw_symbol_token_ref;
 mod reader;
 mod stream_reader;
+mod symbol;
 mod symbol_table;
 mod system_reader;
 mod writer;
 
-pub use binary::raw_binary_reader::RawBinaryReader;
 pub use data_source::IonDataSource;
-pub use raw_reader::RawStreamItem;
 pub use raw_symbol_token::RawSymbolToken;
-pub use reader::Reader;
-pub use reader::StreamItem;
-pub use stream_reader::StreamReader;
-pub use symbol_table::{Symbol, SymbolTable};
-pub use system_reader::{SystemReader, SystemStreamItem};
+pub use raw_symbol_token_ref::RawSymbolTokenRef;
+
+pub use symbol::Symbol;
+pub use symbol_table::SymbolTable;
 
 pub use types::decimal::Decimal;
 pub use types::integer::Integer;
@@ -50,6 +48,14 @@ pub use types::IonType;
 pub use binary::binary_writer::BinaryWriter;
 pub use text::text_writer::TextWriter;
 pub use writer::Writer;
+
+pub use binary::raw_binary_reader::RawBinaryReader;
+pub use raw_reader::{RawReader, RawStreamItem};
+pub use reader::Reader;
+pub use reader::StreamItem;
+pub use stream_reader::StreamReader;
+pub use system_reader::{SystemReader, SystemStreamItem};
+pub use text::raw_text_writer::RawTextWriter;
 
 pub use result::IonError;
 pub use result::IonResult;
