+++ src/reader.rs
@@ -8,7 +8,8 @@ use crate::raw_reader::{RawReader, RawStreamItem};
 use crate::raw_symbol_token::RawSymbolToken;
 use crate::result::{decoding_error, decoding_error_raw, IonResult};
 use crate::stream_reader::StreamReader;
-use crate::symbol_table::{Symbol, SymbolTable};
+use crate::symbol::Symbol;
+use crate::symbol_table::SymbolTable;
 use crate::types::decimal::Decimal;
 use crate::types::integer::Integer;
 use crate::types::timestamp::Timestamp;
